+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/
 
 - Voting system backend
 - Project collaboration backend
+- User profile and dashboard backend
 
 ### Fixed